// apps/web/src/shared/api/generated/model/readingQueueControllerGetQueueParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */
import type { ReadingQueueControllerGetQueueAgeCategoryItem } from "./readingQueueControllerGetQueueAgeCategoryItem";
import type { ReadingQueueControllerGetQueueBookType } from "./readingQueueControllerGetQueueBookType";
import type { ReadingQueueControllerGetQueueFormatItem } from "./readingQueueControllerGetQueueFormatItem";
import type { ReadingQueueControllerGetQueueLanguageItem } from "./readingQueueControllerGetQueueLanguageItem";
import type { ReadingQueueControllerGetQueueOwnerItem } from "./readingQueueControllerGetQueueOwnerItem";
import type { ReadingQueueControllerGetQueuePriorityItem } from "./readingQueueControllerGetQueuePriorityItem";
import type { ReadingQueueControllerGetQueueStatusItem } from "./readingQueueControllerGetQueueStatusItem";

export type ReadingQueueControllerGetQueueParams = {
  /**
   * @maxItems 100
   */
  ageCategory?: ReadingQueueControllerGetQueueAgeCategoryItem[];
  /**
   * @maxItems 100
   * @items.pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  author?: string[];
  bookType?: ReadingQueueControllerGetQueueBookType;
  /**
   * @maxItems 100
   */
  format?: ReadingQueueControllerGetQueueFormatItem[];
  /**
   * @maxItems 100
   * @items.minLength 1
   * @items.maxLength 64
   */
  genre?: string[];
  hasCover?: string;
  /**
   * @maxItems 100
   */
  language?: ReadingQueueControllerGetQueueLanguageItem[];
  /**
   * @maxItems 100
   */
  owner?: ReadingQueueControllerGetQueueOwnerItem[];
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  pagesMax?: number;
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  pagesMin?: number;
  /**
   * @maxItems 100
   */
  priority?: ReadingQueueControllerGetQueuePriorityItem[];
  /**
   * @maxItems 100
   * @items.pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  publisher?: string[];
  /**
   * @maxLength 200
   */
  q?: string;
  /**
   * @minimum 0.5
   * @maximum 10
   */
  ratingMax?: number;
  /**
   * @minimum 0.5
   * @maximum 10
   */
  ratingMin?: number;
  /**
   * @maxItems 100
   */
  status?: ReadingQueueControllerGetQueueStatusItem[];
  /**
   * @maxItems 100
   * @items.pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
   */
  tag?: string[];
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  yearMax?: number;
  /**
   * @minimum -9007199254740991
   * @maximum 9007199254740991
   */
  yearMin?: number;
};